+++ src/main/java/gov/nasa/pds/supp/Constants.java
@@ -1,0 +1,11 @@
+package gov.nasa.pds.supp;
+
+/**
+ * Constants used by different classes
+ * 
+ * @author karpenko
+ */
+public interface Constants
+{
+    public static final String DOI_FIELD = "ops.Identifiers/ops.doi";
+}
